### 2024-10-31 #7377 @JZauner
Dear JOSS Team,

We internally discussed changing the license of LightLogR to a more permissive one, i.e. to MIT instead of GPL. As the licence is documented in the paper, it would be great if we can change this in the paper and the paper branch and merge it to the main branch, so that the license change does not happen after publication.
Would that be acceptable to JOSS? If so, please give a short indication and I will update the branch's license and the relevant part in the manuscript.
Best, Johannes
