### 2024-10-31 #7377 @rwegener2
Hi @JZauner! Not a problem. Feel free to change the license, just merge into the branch.
