### 2024-11-07 #7377 @rwegener2
👋🏻 @hantswilliams and @epiben, would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
