### 2024-11-26 #7377 @cansavvy
> 👋🏻 @cansavvy or @amalthomas111, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I won't be able to have the bandwidth for this immediately but I am interested. What's the deadline?
