### 2024-12-10 c48f7f9
Update license to MIT
