### 2024-12-10 1b47050
Update in paper text to reflect license change to MIT
