+++ paper/paper.md
@@ -52,7 +52,7 @@ affiliations:
 
 # Summary
 
-Light plays an important role in human health and well-being, which necessitates the study of the effects of personal light exposure in real-world settings, measured by means of wearable devices. A growing number of studies incorporate these kinds of data to assess associations between light and health outcomes. Yet with few or missing standards, guidelines, and frameworks, setting up measurements, analysing the data, and comparing outcomes between studies is challenging, especially considering the significantly more complex time series data from wearable light loggers compared to controlled stimuli used in laboratory studies. In this paper, we introduce `LightLogR`, a novel resource to facilitate these research efforts in the form of an open-source, GPL-3.0-licenced software package for the statistical software R. As part of a developing software ecosystem, `LightLogR` is built with common challenges of current and future datasets in mind. The package standardizes many tasks for importing and processing personal light exposure data, provides quick as well as detailed insights into the datasets through summary and visualization tools, and incorporates major metrics commonly used in the field (61 metrics across 17 metric families), while embracing an inherently hierarchical, participant-based data structure.
+Light plays an important role in human health and well-being, which necessitates the study of the effects of personal light exposure in real-world settings, measured by means of wearable devices. A growing number of studies incorporate these kinds of data to assess associations between light and health outcomes. Yet with few or missing standards, guidelines, and frameworks, setting up measurements, analysing the data, and comparing outcomes between studies is challenging, especially considering the significantly more complex time series data from wearable light loggers compared to controlled stimuli used in laboratory studies. In this paper, we introduce `LightLogR`, a novel resource to facilitate these research efforts in the form of an open-source, permissively MIT-licenced software package for the statistical software R. As part of a developing software ecosystem, `LightLogR` is built with common challenges of current and future datasets in mind. The package standardizes many tasks for importing and processing personal light exposure data, provides quick as well as detailed insights into the datasets through summary and visualization tools, and incorporates major metrics commonly used in the field (61 metrics across 17 metric families), while embracing an inherently hierarchical, participant-based data structure.
 
 ![LightLogR logo \label{fig:one}](logo.png){width="25%"}
 
