### 2024-12-17 #7377 @rwegener2
👋🏻 @zhanzmr or @welch16, would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
