### 2024-12-17 #7377 @rwegener2
Wow, thanks for such a speedy reply both @welch16 and @zhanzmr!

As the first to reply I've added @welch16 to the list of reviewers. @zhanzmr we would love to have your expertise as a third reviewer, if you would like. I'm sure your input would improve the quality of this submission.

@zhanzmr I will note that JOSS only requires two reviewers and adding you as a third reviewer is not a JOSS requirement. As such, if you would like to pass on this review that is totally fine and I'm sure JOSS will call on your expertise for a future submission.

Either way is excellent for me - @zhanzmr just let me know if you'd like to be added as a reviewer to this submission, or if you prefer hold off and wait to review a different JOSS submission. Thanks!
