### 2024-12-18 #7601 @editorialbot
[REVIEW]: LightLogR: Reproducible analysis of personal light exposure data
**Submitting author:** @JZauner (<a href="http://orcid.org/0000-0003-2171-4566">Johannes Zauner</a>)
**Repository:** https://github.com/tscnlab/LightLogR
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.5.3
