### 2024-12-18 #7601 @rwegener2
👋🏼 @JZauner @cansavvy, @welch16 this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
