### 2024-12-20 #7601 @cansavvy
@JZauner This is a very cool software! I really appreciate the thoroughness here. Enjoyed taking a look at the code base and paper! I don't work with light loggers so the application is not something I have experience with but can certainly understand the need and appreciate software that attempts to standardize data!

So only have the minorest of comments and questions here:

- Your tutorial articles are great! Very nice. So often I see developers not put much care into those even though that's what really increases your usability.
- There doesn't appear to be a contributing guide or information for contributors anywhere. I think GitHub has a template one you can borrow and adjust if you go to Settings > Community Standards and click the add button next to "CONTRIBUTING.md"
- I ran through the steps in order on the README and many of the steps worked quite seamlessly! But it did tell me `LLdata was not found` and it was unclear in the code where that should be pulled from?
- One of your sentences in the summary in the paper was a very long and a bit hard to keep track of so breaking that up into two might be helpful.
