### 2024-12-29 #7601 @welch16
@JZauner, I agree with @cansavvy that this software is great. In addition to @cansavvy comments, I would like to add:

- Besides adding a contribution guide, adding a link to that file in the `Import` vignette could be useful. In the current version, it states that they can contact you to add it, but could be easier to do through GitHub issues or something.
- I tried to run the vignettes after downloading the current GitHub version, and I ran into some minor issues with the relative location of files. Perhaps using the `here` R package could make the process of manually running the vignettes easier.
- A minor suggestion would be to follow the tidyverse style guide for pipes (<https://style.tidyverse.org/pipes.html>).

Overall, I think it is a great package that would be very useful for the community. Good work!
