### 2025-01-06 #7601 @rwegener2
Thanks @cansavvy and @welch16 for the thoughtful review comments!

@JZauner, please follow up in this review issue once you have completed the review feedback.
