### 2025-01-22 46f298b
rewrote the section on gg_overview so that it does not throw an error
