+++ README.md
@@ -83,15 +83,23 @@ At present, these are the devices we support in LightLogR:
 
 - DeLux
 
+- GENEActiv_GGIR
+
 - Kronowise
 
 - LiDo
 
 - LightWatcher
 
+- LIMO
+
 - LYS
 
+- MotionWatch8
+
 - nanoLambda
+
+- OcuWEAR
 
 - Speccy
 
@@ -120,26 +128,26 @@ section](https://tscnlab.github.io/LightLogR/reference/index.html#metrics).
 There is also an overview article on how to use
 [Metrics](https://tscnlab.github.io/LightLogR/articles/Metrics.html).
 
-| Metric Family                        | Submetrics     | Note                 | Documentation                                                                                    |
-|--------------------------------------|----------------|----------------------|--------------------------------------------------------------------------------------------------|
-| Barroso                              | 7              |                      | `barroso_lighting_metrics()`                                                                     |
-| Bright-dark period                   | 4x2            | bright / dark        | `bright_dark_period()`                                                                           |
-| Centroid of light exposure           | 1              |                      | `centroidLE()`                                                                                   |
-| Disparity index                      | 1              |                      | `disparity_index()`                                                                              |
-| Duration above threshold             | 3              | above, below, within | `duration_above_threshold()`                                                                     |
-| Exponential moving average (EMA)     | 1              |                      | `exponential_moving_average()`                                                                   |
-| Frequency crossing threshold         | 1              |                      | `frequency_crossing_threshold()`                                                                 |
-| Intradaily Variance (IV)             | 1              |                      | `intradaily_variability()`                                                                       |
-| Interdaily Stability (IS)            | 1              |                      | `interdaily_stability()`                                                                         |
-| Midpoint CE (Cumulative Exposure)    | 1              |                      | `midpointCE()`                                                                                   |
-| nvRC (Non-visual circadian response) | 4              |                      | `nvRC()`, `nvRC_circadianDisturbance()`, `nvRC_circadianBias()`, `nvRC_relativeAmplitudeError()` |
-| nvRD (Non-visual direct response)    | 2              |                      | `nvRD()`, `nvRD_cumulative_response()`                                                           |
-| Period above threshold               | 3              | above, below, within | `period_above_threshold()`                                                                       |
-| Pulses above threshold               | 7x3            | above, below, within | `pulses_above_threshold()`                                                                       |
-| Threshold for duration               | 2              | above, below         | `threshold_for_duration()`                                                                       |
-| Timing above threshold               | 3              | above, below, within | `timing_above_threshold()`                                                                       |
-| **Total:**                           |                |                      |                                                                                                  |
-| **17 families**                      | **61 metrics** |                      |                                                                                                  |
+| Metric Family | Submetrics | Note | Documentation |
+|----|----|----|----|
+| Barroso | 7 |  | `barroso_lighting_metrics()` |
+| Bright-dark period | 4x2 | bright / dark | `bright_dark_period()` |
+| Centroid of light exposure | 1 |  | `centroidLE()` |
+| Disparity index | 1 |  | `disparity_index()` |
[…]
