+++ README.Rmd
@@ -142,15 +142,16 @@ dataset %>% ungroup() %>% select(Datetime, lux, kelvin, MEDI) %>%
 
 For more complex data, there is the useful `gg_overview()` function to get an immediate grasp of your data. It was automatically called during import (set `auto.plot = FALSE` to suppress this), but really shines for datasets with multiple participants. It also indicates where data is missing, based on the measurement epochs found in the data.
 
-```         
-LLdata %>% gg_overview()
-```
-
 ::: {style="text-align:center"}
-![](man/figures/gg_overview2.png){width="60%"}
+![Example for `gg_overview()` from a large data collection effort over many months](man/figures/gg_overview2.png){width="60%"}
 :::
 
-*note:* the above example can not be executed through copy and paste, as it requires a large dataset not included in the package. It is available, however, in the article on [Import & cleaning](https://tscnlab.github.io/LightLogR/articles/Import.html).
+*note:* the above example image requires a large dataset not included in the package. It is available, however, in the article on [Import & cleaning](https://tscnlab.github.io/LightLogR/articles/Import.html).
+
+```
+#example code, on how to use gg_overview():
+dataset %>% gg_overview()
+```
 
 ### Visualize
 
