### 2025-01-22 713fadf
added and populated a CONTRIBUTING.md file
