+++ .github/CONTRIBUTING.md
@@ -1,0 +1,152 @@
+<!-- omit in toc -->
+# Contributing to LightLogR
+
+First off, thanks for taking the time to contribute! ❤️
+
+All types of contributions are encouraged and valued. See the [Table of Contents](#table-of-contents) for different ways to help and details about how this project handles them. Please make sure to read the relevant section before making your contribution. It will make it a lot easier for us maintainers and smooth out the experience for all involved. The community looks forward to your contributions. 🎉
+
+> And if you like the project, but just don't have time to contribute, that's fine. There are other easy ways to support the project and show your appreciation, which we would also be very happy about:
+> - Star the project
+> - Tweet about it
+> - Refer this project in your project's readme or publications that make use of the software
+> - Mention the project at local meetups and tell your friends/colleagues
+
+<!-- omit in toc -->
+## Table of Contents
+
+- [Code of Conduct](#code-of-conduct)
+- [I Have a Question](#i-have-a-question)
+- [Why is my light logger, dosimeter, or other wearable device not supported by LightLogR](#i-want-my-logger)
+- [I want News about LightLogR](#i-want-news)
+- [I Want To Contribute](#i-want-to-contribute)
+- [Reporting Bugs](#reporting-bugs)
+- [Suggesting Enhancements](#suggesting-enhancements)
+- [Your First Code Contribution](#your-first-code-contribution)
+- [Improving The Documentation](#improving-the-documentation)
+- [Styleguides](#styleguides)
+- [Commit Messages](#commit-messages)
+- [Join The Project Team](#join-the-project-team)
+
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by the
+[LightLogR Code of Conduct](https://github.com/tscnlab/LightLogR/blob//CODE_OF_CONDUCT.md).
+By participating, you are expected to uphold this code. Please report unacceptable behavior
+to the [lead developer](mailto:johannes.zauner@tum.de).
+
+
+## I Have a Question
+
+> If you want to ask a question, we assume that you have read the available [Documentation](https://tscnlab.github.io/LightLogR/).
+
+Before you ask a question, it is best to search for existing [Issues](https://github.com/tscnlab/LightLogR/issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in this issue. It is also advisable to search the internet for answers first.
+
+If you then still feel the need to ask a question and need clarification, we recommend the following:
+
+- Open an [Issue](https://github.com/tscnlab/LightLogR/issues/new).
+- Provide as much context as you can about what you're running into.
+- Provide project and platform versions, depending on what seems relevant.
+
+We will then take care of the issue as soon as possible.
+
+## Why is my light logger, dosimeter, or other wearable device not supported by LightLogR
+
+If you are using a device that is currently not supported, please contact us. 
+You can find a list of supported devices in the [LightLogR documentation about imports](https://tscnlab.github.io/LightLogR/reference/import_Dataset.html#devices).
+
+We are always looking to expand the range of supported devices. The easiest and most trackable way to get in contact is by opening a new issue on our [Github repository](https://github.com/tscnlab/LightLogR/issues). Please also provide a sample file of your data, so we can test the import function.
+
+## I want News about LightLogR
+
+If you are interested in the project and want to know more, you can subscribe to the [LightLogR mailing list](https://lists.lrz.de/mailman/listinfo/lightlogr-users).
+
+## I Want To Contribute
+
+<!-- omit in toc -->
+### Legal Notice 
+
+> When contributing to this project, you must agree that you have authored 100% of the content, that you have the necessary rights to the content and that the content you contribute may be provided under the project licence.
+
+### Reporting Bugs
+
+<!-- omit in toc -->
+#### Before Submitting a Bug Report
+
[…]
