### 2025-01-22 a79c9ef
adding Code of Conduct and Contributing information
