### 2025-01-22 28e88a0
added comments on relative paths and made conditional folder generation
