+++ vignettes/articles/Metrics.Rmd
@@ -23,6 +23,7 @@ library(gtsummary)
 We will use data imported and cleaned already in the article [Import & Cleaning](https://tscnlab.github.io/LightLogR/articles/Import.html).
 
 ```{r, import}
+#this assumes the data is in the cleaned_data folder in the working directory
 data <- readRDS("cleaned_data/ll_data.rds")
 ```
 
