+++ vignettes/articles/Visualizations.Rmd
@@ -39,7 +39,9 @@ As can be seen the dataset contains 17 ids with one weeks worth of data each, an
 
 ```{r}
 data %>%
-  filter(!(date(Datetime) %in% c("2023-08-16", "2023-08-17"))) %>% 
+  filter(
+    !(date(Datetime) %in% c("2023-08-16", "2023-08-17"))
+    ) %>% 
   gg_overview()
 ```
 
@@ -47,7 +49,9 @@ Calculating gaps in the data can be computationally expensive for large datasets
 
 ```{r}
 data %>%
-  filter(!(date(Datetime) %in% c("2023-08-16", "2023-08-17"))) %>% 
+  filter(
+    !(date(Datetime) %in% c("2023-08-16", "2023-08-17"))
+    ) %>% 
   gg_overview(gap.data = tibble())
 ```
 
@@ -83,7 +87,8 @@ Showing the days by date is the default behavior of `gg_day()`. It can also be g
 
 ```{r, fig.width=7, fig.height = 10}
 data %>% 
-  gg_day(aes_col = Id, size = 0.5, format.day = "%A") + scale_color_viridis_d()
+  gg_day(aes_col = Id, size = 0.5, format.day = "%A") + 
+  scale_color_viridis_d()
 ```
 
 ## Customizing geoms and miscellanea
@@ -97,12 +102,14 @@ data_subset <-
   aggregate_Datetime(unit = "15 mins") %>% #aggregating to 15 min intervals
   filter_Datetime(length = "3 days", full.day = TRUE) #restricting their length to 3 days
 
-data_subset %>%  gg_day(aes_col = Id)
+data_subset %>% 
+  gg_day(aes_col = Id)
 ```
 
 Now we can use a different geom.
 ```{r}
-data_subset %>%  gg_day(aes_col = Id, geom = "line")
+data_subset %>%
+  gg_day(aes_col = Id, geom = "line")
 ```
 
 Also a ribbon is possible.
@@ -121,7 +128,8 @@ data %>%
   filter(Id %in% c(205, 216, 219)) %>% #choosing 2 ids
   aggregate_Datetime(unit = "15 mins") #aggregating to 15 min intervals
 
-data_subset2 %>%  gg_days()
+data_subset2 %>%
+  gg_days()
 ```
 
 By default, `gg_days()` will always plot full days. Let us strip one participant of data for three days.
@@ -132,7 +140,8 @@ data_subset2 %>%
   filter(!(Id == 205 & 
              date(Datetime) %in% c("2023-08-28", "2023-08-29", "2023-08-30")))
 
-data_subset3 %>%  gg_days()
+data_subset3 %>%
+  gg_days()
 ```
 
 You can see the plots are misaligned in their facets. We can correct for that by providing an exact number of days to the `x.axis.limits` argument. `Datetime_limits()` is a helper function from `LightLogR` and the documentation reveals more about its arguments.
@@ -143,7 +152,6 @@ data_subset3 %>%
     x.axis.limits = 
       \(x) Datetime_limits(x, length = ddays(7), midnight.rollover = TRUE)
     )
-
 ```
 
 `gg_days()` has all of the customization options of `gg_day()`. Here we will customize the plot for a ribbon, different naming and breaks on the datetime axis.
@@ -159,7 +167,6 @@ data_subset3 %>%
     x.axis.format = "%H",
     ) +
   guides(color = "none", fill = "none")
-
 ```
 
 # gg_doubleplot()
@@ -171,7 +178,9 @@ data_subset3 %>%
[…]
