+++ vignettes/articles/Day.Rmd
@@ -63,8 +63,10 @@ As you can see, the import is accompanied by a (hopefully) helpful message about
 Because we have no missing values that we would have to deal with first, this dataset is already good to go. If you, e.g., want to know the range of melanopic EDI (a measure of stimulus strength for the nonvisual system) for every day in the dataset, you can do that:
 
 ```{r}
-dataset.LL %>% group_by(Date = as_date(Datetime)) %>% 
-  summarize(range.MEDI = range(MEDI) %>% str_flatten(" - ")) %>% gt()
+dataset.LL %>% 
+  group_by(Date = as_date(Datetime)) %>% 
+  summarize(range.MEDI = range(MEDI) %>% str_flatten(" - ")) %>% 
+  gt()
 ```
 
 Same goes for visualization - it is always helpful to get a good look at data immediately after import. The `gg_day()` function creates a simple `ggplot` of the data, stacked vertically by Days. The function needs very little input beyond the dataset (in fact, it would even work without the `size` input, which just makes the default point size smaller, and the `interactive` command sends the output to `plotly` to facilitate data exploration). `gg_day()` features a lot of flexibility, and can be adapted and extended to fit various needs, as we will see shortly.
@@ -102,7 +104,9 @@ dataset.sleep <-
                       dec = ",",
                       tz = tz)
 
-dataset.sleep %>% head() %>% gt()
+dataset.sleep %>% 
+  head() %>% 
+  gt()
 ```
 
 Now that we have imported all of our data, we need to combine it sensibly, which we will get to in the next section.
@@ -123,13 +127,16 @@ With the `data2reference()` function, we will create this reference. This functi
 
 ```{r}
 dataset.LL <- 
-  dataset.LL %>% data2reference(Reference.data = dataset.env, across.id = TRUE)
+  dataset.LL %>% 
+    data2reference(Reference.data = dataset.env, across.id = TRUE)
 
 dataset.LL <- 
-  dataset.LL %>% select(Id, Datetime, MEDI, Reference)
+  dataset.LL %>% 
+    select(Id, Datetime, MEDI, Reference)
 
-dataset.LL %>% head() %>% gt()
-
+dataset.LL %>% 
+  head() %>% 
+  gt()
 ```
 
 For the sake of this example, we have also removed unnecessary data columns, which makes the further code examples simpler. We can already see in the table above that the reference at the start of the measurements is quite a bit higher than the luminous exposure at the participant´s light logger. We also see that the same reference value is applied to three participant values. This mirrors the fact that for every three measurements taken with the participant device, one measurement epoch for the environmental sensor passes.
@@ -138,7 +145,8 @@ To visualize this newly reached reference data, we can easily extend `gg_day()` 
 
 ```{r, fig.retina=2, fig.width=7, fig.height = 10}
 dataset.LL %>% 
-  gg_day(size = 0.25) + geom_line(aes(y=Reference), lty = 2, col = "red")
+  gg_day(size = 0.25) + 
+  geom_line(aes(y=Reference), lty = 2, col = "red")
 ```
 
[…]

[…]
