### 2025-01-22 246e2bb
update to News about the review
