### 2025-01-23 fc4873e
replaced the magrittr pipe with the native pipe
