+++ vignettes/articles/Metrics.Rmd
@@ -18,6 +18,8 @@ library(gt)
 library(gtsummary)
 ```
 
+> Please note that this article uses the base pipe operator `|>`. You need an R version equal to or greater than 4.1.0 to use it. If you are using an older version, you can replace it with the `magrittr` pipe operator `%>%`.
+
 # Importing Data
 
 We will use data imported and cleaned already in the article [Import & Cleaning](https://tscnlab.github.io/LightLogR/articles/Import.html).
@@ -30,7 +32,7 @@ data <- readRDS("cleaned_data/ll_data.rds")
 As can be seen by using `gg_overview()`, the dataset contains 17 ids with one weeks worth of data each, and one to three participants per week.
 
 ```{r, overview}
-data %>% gg_overview()
+data |> gg_overview()
 ```
 
 # Metric principles
@@ -53,10 +55,10 @@ All metric functions are by default agnostic to the type of data. They require v
 
 ```{r, Id 201}
 data_Id201 <- 
-  data %>% 
+  data |> 
     filter(Id == 201 & date(Datetime) == "2023-08-15")
 
-data_Id201 %>% 
+data_Id201 |> 
   gg_day()
 ```
 
@@ -103,7 +105,7 @@ bright_dark_period(
   Light.vector = data_Id201$MEDI,
   Time.vector = data_Id201$Datetime,
   as.df = TRUE
-) %>% 
+) |> 
   gt()
 ```
 
@@ -120,7 +122,7 @@ bright_dark_period(
   timespan = "10 hours"
 )
 
-M10_wrong %>% gt()
+M10_wrong |> gt()
 ```
 
 We also see that this makes little sense, if we visualize this portion. The yellow color indicates the darkest 10-hour period of the day.
@@ -129,7 +131,7 @@ We also see that this makes little sense, if we visualize this portion. The yell
 Onset <- M10_wrong$darkest_10h_onset
 Offset <- M10_wrong$darkest_10h_offset
 
-data_Id201 %>% 
+data_Id201 |> 
   gg_day(aes_col = Datetime >= Onset & Datetime <= Offset) +
   guides(color = "none")
 
@@ -148,7 +150,7 @@ bright_dark_period(
   loop = TRUE
 )
 
-M10 %>% gt()
+M10 |> gt()
 ```
 
 This is more plausible, and can also be visualized easily.
@@ -157,7 +159,7 @@ This is more plausible, and can also be visualized easily.
 Onset <- M10$darkest_10h_onset
 Offset <- M10$darkest_10h_offset
 
-data_Id201 %>% 
+data_Id201 |> 
   gg_day(aes_col = Datetime >= Onset | Datetime <= Offset) +
   guides(color = "none")
 ```
@@ -172,7 +174,7 @@ As we only want to calculate metrics for days with full data, we will exclude Mo
 
 ```{r}
 data <- 
-  data %>% 
+  data |> 
   dplyr::filter(weekdays(Datetime) != "Monday")
 ```
 
@@ -181,14 +183,14 @@ data <-
 The `dplyr::summarize()` function is used to calculate metrics for each group of data. In the following example, we will calculate Interdaily Stability (IS) for all participants in the data set, giving us the variability of the 24h light exposure patterns across the full 6 days of data compared to their average, ranging between 0 (Gaussian noise) and 1 (Perfect stability). For brevity, only the first 6 Ids will be shown.
 
 ```{r}
-data %>% 
+data |> 
   summarize(
     IS = interdaily_stability(
       Light.vector = MEDI,
       Datetime.vector = Datetime
     )
-  ) %>% 
-  head() %>% 
+  ) |> 
+  head() |> 
   gt() 
 ```
 
@@ -197,15 +199,15 @@ data %>%
[…]
