+++ vignettes/articles/Visualizations.Rmd
@@ -18,6 +18,8 @@ library(tidyverse)
 library(patchwork)
 ```
 
+> Please note that this article uses the base pipe operator `|>`. You need an R version equal to or greater than 4.1.0 to use it. If you are using an older version, you can replace it with the `magrittr` pipe operator `%>%`.
+
 # Importing Data
 
 We will use data imported and cleaned already in the article [Import & Cleaning](https://tscnlab.github.io/LightLogR/articles/Import.html).
@@ -32,26 +34,26 @@ data <- readRDS("cleaned_data/ll_data.rds")
 `gg_overview()` provides a glance at *when* data is available for each Id. Let's call it on our dataset.
 
 ```{r, overview}
-data %>% gg_overview()
+data |> gg_overview()
 ```
 
 As can be seen the dataset contains 17 ids with one weeks worth of data each, and one to three participants per week. `gg_overview()` will by default test whether there are gaps in the data and will show them as grey bars, as well as a message in the lower right corner. Let us force this behavior in our dataset by removing two days.
 
 ```{r}
-data %>%
+data |>
   filter(
     !(date(Datetime) %in% c("2023-08-16", "2023-08-17"))
-    ) %>% 
+    ) |> 
   gg_overview()
 ```
 
 Calculating gaps in the data can be computationally expensive for large datasets with small epochs. If you just require an overview of the data without being concerned about gaps, you can provide an empty `tibble::tibble()` to the `gap.data` argument. This will skip the gap calculation and speed up the graph generation.
 
 ```{r}
-data %>%
+data |>
   filter(
     !(date(Datetime) %in% c("2023-08-16", "2023-08-17"))
-    ) %>% 
+    ) |> 
   gg_overview(gap.data = tibble())
 ```
 
@@ -63,8 +65,8 @@ data %>%
 `gg_day()` compares days within a dataset. By default it will use the `date`. Let's call it on a subset of our data. To distinguish between different Ids, we can set the `aes_col` argument to `Id`.
 
 ```{r, fig.width=7, fig.height = 10}
-data %>% 
-  filter(Id %in% c(205, 206)) %>% 
+data |> 
+  filter(Id %in% c(205, 206)) |> 
   gg_day(aes_col = Id, size = 0.5)
 ```
 
@@ -73,8 +75,8 @@ data %>%
 Note that each day is represented by its own facet, which is named after the date. We can give each Id its own facet by using the `ggplot2::facet_wrap()` function. The `Day.data` column is produced by `gg_day()` and contains the structure of the daily facets. It has to be used by `facet_wrap()` to ensure that the facets are shown correctly. We also reduce the breaks on the x-axis to avoid overlap at 00:00.
 
 ```{r, fig.width=7, fig.height = 10}
-data %>% 
-  filter(Id %in% c(205, 206)) %>% 
+data |> 
+  filter(Id %in% c(205, 206)) |> 
   gg_day(aes_col = Id, size = 0.5, 
          x.axis.breaks = hms::hms(hours = c(0, 6, 12, 18))) + 
   guides(color = "none") +
@@ -86,7 +88,7 @@ data %>%
[…]

[…]
