### 2025-01-23 685131f
replaced the magrittr pipe with the native pipe
