+++ vignettes/articles/Day.Rmd
@@ -65,16 +65,18 @@ As you can see, the import is accompanied by a (hopefully) helpful message about
 Because we have no missing values that we would have to deal with first, this dataset is already good to go. If you, e.g., want to know the range of melanopic EDI (a measure of stimulus strength for the nonvisual system) for every day in the dataset, you can do that:
 
 ```{r}
-dataset.LL %>% 
-  group_by(Date = as_date(Datetime)) %>% 
-  summarize(range.MEDI = range(MEDI) %>% str_flatten(" - ")) %>% 
+dataset.LL |> 
+  group_by(Date = as_date(Datetime)) |> 
+  summarize(
+    range.MEDI = range(MEDI) |> str_flatten(" - ")
+    ) |> 
   gt()
 ```
 
 Same goes for visualization - it is always helpful to get a good look at data immediately after import. The `gg_day()` function creates a simple `ggplot` of the data, stacked vertically by Days. The function needs very little input beyond the dataset (in fact, it would even work without the `size` input, which just makes the default point size smaller, and the `interactive` command sends the output to `plotly` to facilitate data exploration). `gg_day()` features a lot of flexibility, and can be adapted and extended to fit various needs, as we will see shortly.
 
 ```{r, fig.retina=2, fig.width=7, fig.height = 14}
-dataset.LL %>% gg_day(size = 0.25, interactive = TRUE)
+dataset.LL |> gg_day(size = 0.25, interactive = TRUE)
 ```
 
 We can already see some patterns and features in the luminous exposure across the days. In general, the participant seems to have woken (or at least started wearing the light logger) after 9:00 and went to bed (or, again, stopped wearing the device) at around 23:00.
@@ -106,8 +108,8 @@ dataset.sleep <-
                       dec = ",",
                       tz = tz)
 
-dataset.sleep %>% 
-  head() %>% 
+dataset.sleep |> 
+  head() |> 
   gt()
 ```
 
@@ -129,15 +131,15 @@ With the `data2reference()` function, we will create this reference. This functi
 
 ```{r}
 dataset.LL <- 
-  dataset.LL %>% 
+  dataset.LL |> 
     data2reference(Reference.data = dataset.env, across.id = TRUE)
 
 dataset.LL <- 
-  dataset.LL %>% 
+  dataset.LL |> 
     select(Id, Datetime, MEDI, Reference)
 
-dataset.LL %>% 
-  head() %>% 
+dataset.LL |> 
+  head() |> 
   gt()
 ```
 
@@ -146,7 +148,7 @@ For the sake of this example, we have also removed unnecessary data columns, whi
 To visualize this newly reached reference data, we can easily extend `gg_day()` with a dashed red reference line. Keep in mind that this visualization is still exploratory, so we are not investing heavily in styling.
 
 ```{r, fig.retina=2, fig.width=7, fig.height = 10}
-dataset.LL %>% 
+dataset.LL |> 
   gg_day(size = 0.25) + 
   geom_line(aes(y=Reference), lty = 2, col = "red")
 ```
@@ -173,11 +175,11 @@ In our case the first entry is sleep, so we can safely assume, that on that day 
 
 ```{r}
 dataset.sleep <- 
-  dataset.sleep %>% 
+  dataset.sleep |> 
[…]
