+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   articles/Import: Import.html
   articles/Metrics: Metrics.html
   articles/Visualizations: Visualizations.html
-last_built: 2024-12-10T11:52Z
+last_built: 2025-01-23T10:44Z
 urls:
   reference: https://tscnlab.github.io/LightLogR/reference
   article: https://tscnlab.github.io/LightLogR/articles
