### 2025-01-23 #7601 @JZauner
Thank you very much, @cansavvy and @welch16, for your kind words and constructive suggestions, and to @rwegener2 for coordinating everything! We strive to create a well-developed and user-accessible package, which is why we emphasized tutorials. We believe the suggested changes have made the package more robust and user-friendly.

We have implemented the following changes, which are live now as [LightLogR 0.4.3](https://tscnlab.github.io/LightLogR/). They address all suggestions the reviewers made.

- Added a [CONTRIBUTING.md](https://tscnlab.github.io/LightLogR/CONTRIBUTING.html) file as well as a [CODE_OF_CONDUCT.md](https://tscnlab.github.io/LightLogR/CODE_OF_CONDUCT.html) file and referenced it both in the [README](https://tscnlab.github.io/LightLogR/index.html#i-want-to-contribute), and the [Import vignette](https://tscnlab.github.io/LightLogR/articles/Import.html#what-if-my-device-is-not-listed)
- [Broken up the paper summary](https://github.com/tscnlab/LightLogR/pull/43/commits/198cad4641d759c803a0a2ffe3528088c34f1181) into smaller sentences (note: this is in the `main` branch, not the `paper` branch)
- Rewrote all [vignettes/articles](https://tscnlab.github.io/LightLogR/articles/) with the tidyverse style guide for pipes, including using the native pipe
- Rewrote the [README section](https://github.com/tscnlab/LightLogR/pull/43/commits/46f298b3cdc1acef6ad72ad6429199b6897f72a7) that referenced a non-existing object `LLdata` so that it is unambiguous, how to access that data
- Slight changes to the [vignettes/articles](https://tscnlab.github.io/LightLogR/articles/) in sections with relative paths. I ran into problems using the `here` package (because it would have required me to rephrase paths according to the package structure, which I don't think is helpful for the user/viewer). Instead, I added comments to make clear where data is stored or accessed relative to the working directory, and also added a code line to conditionally create necessary folders, should they not exist already. This should eliminate most if not all troubles when rerunning them locally.
