+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   articles/Import: Import.html
   articles/Metrics: Metrics.html
   articles/Visualizations: Visualizations.html
-last_built: 2025-01-23T10:44Z
+last_built: 2025-01-23T11:21Z
 urls:
   reference: https://tscnlab.github.io/LightLogR/reference
   article: https://tscnlab.github.io/LightLogR/articles
