### 2025-02-03 2bae2e6
new function photoperiod, with tests
