### 2025-02-03 b86e39d
extended the output, example in documentation
