+++ man/photoperiod.Rd
@@ -1,0 +1,51 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/photoperiod.R
+\name{photoperiod}
+\alias{photoperiod}
+\title{Calculate boundary times of the photoperiod}
+\usage{
+photoperiod(coordinates, dates, tz, solarDep = 6)
+}
+\arguments{
+\item{coordinates}{A two element numeric vector representing the latitude and
+longitude of the location. \emph{Important note: \strong{Latitude} is the first
+element and \strong{Longitude} is the second element.}}
+
+\item{dates}{A date of format \code{Date}, or coercible to \code{Date} through
+\code{\link[lubridate:as_date]{lubridate::as_date()}}}
+
+\item{tz}{Timezone of the data. Expects a \code{character}. You can look up the
+supported timezones with \code{\link[=OlsonNames]{OlsonNames()}}.}
+
+\item{solarDep}{A numerical value representing the solar depression angle
+between 90 and -90. This means a value of 6 equals \strong{-6} degrees above the
+horizon. Default is 6, equalling \verb{Civil dawn/dusk}. Other common values are
+12 degrees for \verb{Nautical dawn/dusk}, 18 degrees for \verb{Astronomical dawn/dusk}, and 0 degrees for \code{Sunrise/Sunset}. Note that the output
+columns will always be named \code{dawn} and \code{dusk}, regardless of the
+\code{solarDep} value.}
+}
+\value{
+A \code{tibble} with the calculated times of dawn and dusk for the given
+location and date. The tibble contains the following columns:
+\itemize{
+\item \code{date} with the date of the calculation, stored as class \code{Date}
+\item \code{tz} with the timezone of the output, stored as class \code{character}
+\item \code{lat} and \code{lon} with the latitude and longitude of the location, stored
+as class \code{numeric}
+\item \code{solar.angle} with the negative solar depression angle, i.e. the sun
+elevation above the horizon. stored as class \code{numeric}
+\item \code{dawn} and \code{dusk} with the calculated datetimes, stored as class \code{POSIXct}
+}
+}
+\description{
+Calculate boundary times of the photoperiod
+}
+\examples{
+  coordinates <- c(47.018711, 12.34256)
+  dates <- c("2023-06-01", "2025-08-23")
+  tz <- "Europe/Zurich"
+  #civil dawn/dusk
+  photoperiod(coordinates, dates, tz)
+  #sunrise/sunset
+  photoperiod(coordinates, dates, tz, solarDep = 0)
+}
