+++ NAMESPACE
@@ -50,6 +50,7 @@ export(nvRC_relativeAmplitudeError)
 export(nvRD)
 export(nvRD_cumulative_response)
 export(period_above_threshold)
+export(photoperiod)
 export(pulses_above_threshold)
 export(sc2interval)
 export(sleep_int2Brown)
