### 2025-02-04 #7601 @rwegener2
Lovely. Thanks so much again to @welch16 and @cansavvy for your feedback as reviewers! JOSS wouldn't be possible without you.

@JZauner The next step in the review process is that I will go through the paper for a series of checks. I'll be looking for some information from you midway through. After that I will pass it off the the track editor and chief, who will do the final review and publication.
