### 2025-02-04 #7601 @rwegener2
Hi @JZauner, thanks for a well-written paper! It was a pleasure to read. I noticed one small typo:
* line 50: These function**s** (missing an s)

I also noticed there is something odd going on with the table label and reference rendering. It appears that Table 1 doesn't have a title and in line 54 the table reference is not rendering a number. I wonder if it's because of the footnote? I'm not confident that markdown can handle footnotes within a table, but one thing to try would be to move the footnote outside of the table caption line, something like:

```
| last | line | of | table |

: Devices supported for import in version 0.4.1 \label{tab:one}

[^1]: Available after processing of the data using `GGIR` [@migueles2019].

```
Finally, do you reference Table 1 at any point in the text? I'd suggest you either reference it somewhere or take it out.
