### 2025-02-04 #7601 @JZauner
Hi @rwegener2, thank you for your feedback! I will fix those issues as you suggested. I also sent the proof to my coauthors so that they can do final checks (as mentioned in the post-review checklist). I already got feedback from one and am waiting for feedback from the other. Ideally, I will have the final and (coauthor)approved version the day after tomorrow.
