### 2025-02-04 e51b6f6
Added functions, documentation, and tests for extract_photoperiod and add_photoperiod
