+++ R/photoperiod.R
@@ -1,4 +1,17 @@
-#' Calculate boundary times of the photoperiod
+
+# photoperiod -------------------------------------------------------------
+
+
+#' Calculate photoperiod and boundary times
+#'
+#' A family of functions to extract and add photoperiod information.
+#' [photoperiod()] creates a `tibble` with the calculated times of dawn and dusk
+#' for the given location and date. The function is a convenience wrapper for
+#' [suntools::crepuscule()] to calculate the times of dawn and dusk. By default,
+#' civil dawn and dusk are calculated, but the function can be used to calculate
+#' other times by changing the `solarDep` parameter (e.g., 0 for sunrise/sunset,
+#' 12 for nautical, and 18 for astronomical).
+#'
 #'
 #' @param coordinates A two element numeric vector representing the latitude and
 #'   longitude of the location. *Important note: **Latitude** is the first
@@ -15,25 +28,32 @@
 #'   columns will always be named `dawn` and `dusk`, regardless of the
 #'   `solarDep` value.
 #'
-#' @returns A `tibble` with the calculated times of dawn and dusk for the given
-#'   location and date. The tibble contains the following columns:
+#' @returns [photoperiod()] returns a `tibble` with the calculated times of dawn
+#'   and dusk for the given location and date, with the length equal to the
+#'   `dates` input parameter . The tibble contains the following columns:
 #'  - `date` with the date of the calculation, stored as class `Date`
 #'  - `tz` with the timezone of the output, stored as class `character`
 #'  - `lat` and `lon` with the latitude and longitude of the location, stored
-#'    as class `numeric`
-#'  - `solar.angle` with the negative solar depression angle, i.e. the sun 
-#'    elevation above the horizon. stored as class `numeric`
+#'   as class `numeric`
+#'  - `solar.angle` with the negative solar depression angle, i.e. the sun
+#'   elevation above the horizon. stored as class `numeric`
 #'  - `dawn` and `dusk` with the calculated datetimes, stored as class `POSIXct`
+#'  - `photoperiod` with the calculated photoperiod, stored as class `difftime`.
+#'
 #' @export
 #'
+#' @family photoperiod
+#'
 #' @examples
-#'   coordinates <- c(47.018711, 12.34256)
-#'   dates <- c("2023-06-01", "2025-08-23")
-#'   tz <- "Europe/Zurich"
-#'   #civil dawn/dusk
-#'   photoperiod(coordinates, dates, tz)
-#'   #sunrise/sunset
-#'   photoperiod(coordinates, dates, tz, solarDep = 0)
+#'  #example für Tübingen, Germany
+#'  coordinates <- c(48.521637, 9.057645)
+#'  dates <- c("2023-06-01", "2025-08-23")
+#'  tz <- "Europe/Berlin"
+#'  
+#'  #civil dawn/dusk
+#'  photoperiod(coordinates, dates, tz)
+#'  #sunrise/sunset
+#'  photoperiod(coordinates, dates, tz, solarDep = 0)
 
 photoperiod <- function(coordinates, dates, tz, solarDep = 6) {
   
@@ -90,5 +110,196 @@ photoperiod <- function(coordinates, dates, tz, solarDep = 6) {
                   lat = coordinates[1],
                   lon = coordinates[2],
                   solar.angle = -solarDep,
-                  .before = 1)
+                  .before = 1) |> 
+    dplyr::mutate(photoperiod = dusk - dawn)
 }
+
+
+# extract_photoperiod ------------------------------------------------------
+
+
+#' Calculate photoperiods and their boundary times for a light logger dataset
+#'
[…]
