+++ man/photoperiod.Rd
@@ -2,9 +2,26 @@
 % Please edit documentation in R/photoperiod.R
 \name{photoperiod}
 \alias{photoperiod}
-\title{Calculate boundary times of the photoperiod}
+\alias{extract_photoperiod}
+\alias{add_photoperiod}
+\title{Calculate photoperiod and boundary times}
 \usage{
 photoperiod(coordinates, dates, tz, solarDep = 6)
+
+extract_photoperiod(
+  dataset,
+  coordinates,
+  Datetime.colname = Datetime,
+  solarDep = 6
+)
+
+add_photoperiod(
+  dataset,
+  coordinates,
+  Datetime.colname = Datetime,
+  solarDep = 6,
+  overwrite = FALSE
+)
 }
 \arguments{
 \item{coordinates}{A two element numeric vector representing the latitude and
@@ -23,10 +40,20 @@ horizon. Default is 6, equalling \verb{Civil dawn/dusk}. Other common values are
 12 degrees for \verb{Nautical dawn/dusk}, 18 degrees for \verb{Astronomical dawn/dusk}, and 0 degrees for \code{Sunrise/Sunset}. Note that the output
 columns will always be named \code{dawn} and \code{dusk}, regardless of the
 \code{solarDep} value.}
+
+\item{dataset}{A light logger dataset. Expects a \code{dataframe}. If not imported
+by \link{LightLogR}, take care to choose a sensible variable for the
+\code{Datetime.colname}.}
+
+\item{Datetime.colname}{column name that contains the datetime. Defaults to
+\code{"Datetime"} which is automatically correct for data imported with
+\link{LightLogR}. Expects a \code{symbol}. Needs to be part of the \code{dataset}. Must
+be of type \code{POSIXct}.}
 }
 \value{
-A \code{tibble} with the calculated times of dawn and dusk for the given
-location and date. The tibble contains the following columns:
+\code{\link[=photoperiod]{photoperiod()}} returns a \code{tibble} with the calculated times of dawn
+and dusk for the given location and date, with the length equal to the
+\code{dates} input parameter . The tibble contains the following columns:
 \itemize{
 \item \code{date} with the date of the calculation, stored as class \code{Date}
 \item \code{tz} with the timezone of the output, stored as class \code{character}
@@ -35,17 +62,90 @@ as class \code{numeric}
 \item \code{solar.angle} with the negative solar depression angle, i.e. the sun
 elevation above the horizon. stored as class \code{numeric}
 \item \code{dawn} and \code{dusk} with the calculated datetimes, stored as class \code{POSIXct}
+\item \code{photoperiod} with the calculated photoperiod, stored as class \code{difftime}.
 }
+
+\code{\link[=extract_photoperiod]{extract_photoperiod()}} returns a \code{tibble} of the same structure as
+\code{\link[=photoperiod]{photoperiod()}}, but with a length equal to the number of unique dates in
+the dataset.
+
+The input dataset with the added photoperiod information. The
+information is appended with the following columns: \code{dawn}, \code{dusk},
+\code{photoperiod}, and \code{photoperiod.state}.
 }
 \description{
-Calculate boundary times of the photoperiod
+A family of functions to extract and add photoperiod information.
+\code{\link[=photoperiod]{photoperiod()}} creates a \code{tibble} with the calculated times of dawn and dusk
+for the given location and date. The function is a convenience wrapper for
[…]

[…]
