### 2025-02-04 edbee53
added function, documentation, and tests for solar_noon()
