+++ NAMESPACE
@@ -56,6 +56,7 @@ export(photoperiod)
 export(pulses_above_threshold)
 export(sc2interval)
 export(sleep_int2Brown)
+export(solar_noon)
 export(supported_devices)
 export(symlog_trans)
 export(threshold_for_duration)
