### 2025-02-06 5205109
update to sample.data.enviroment so that it is reproducible and has the correct timezone for photoperiod calculation
