+++ man/photoperiod.Rd
@@ -133,7 +133,6 @@ over the locations.
 #add_photoperiod
 added_photoperiod <- 
  sample.data.environment |> 
- dplyr::mutate(Datetime = lubridate::force_tz(Datetime, tz)) |> 
  add_photoperiod(coordinates)
    
 added_photoperiod |> head()
