### 2025-02-10 477e7b0
inclusion of feedback to paper.md manuscript
