+++ paper/paper.md
@@ -17,7 +17,7 @@ tags:
 - wearable sensors
 - wearables
 - "rstats-package"
-date: "07 October 2024"
+date: "10 February 2025"
 output: pdf_document
 authors:
 - name: Johannes Zauner
@@ -35,7 +35,7 @@ editor_options:
   markdown:
     wrap: 72
 affiliations:
-- name: Technical University of Munich, TUM School of Medicine and Health, Department Health and Sports Sciences, Chronobiology & Health, Munich, Germany
+- name: Technical University of Munich, TUM School of Medicine and Health, Department Health and Sport Sciences, Chronobiology & Health, Munich, Germany
   index: 1
   ror: 02kkvpp62
 - name: Max Planck Institute for Biological Cybernetics, Max Planck Research Group Translational Sensory & Circadian Neuroscience, Tübingen, Germany
@@ -58,9 +58,9 @@ Light plays an important role in human health and well-being, which necessitates
 
 # Statement of need
 
-Personalized luminous exposure data is progressively gaining importance across various domains, including research, occupational affairs, and lifestyle tracking. Data are collected through a proliferating selection of wearable light loggers and dosimeters, varying in size, shape, functionality, and output format [@hartmeyer2023]. Despite or potentially because of numerous use cases, the field still lacks a unified framework for collecting, validating, and analyzing the accumulated data [@hartmeyer2023; @spitschan2022].  This issue increases the time and expertise necessary to handle such data and also compromises the FAIRness (findability, accessibility, interoperability, reusability) [@wilkinson2016] of the results, especially for meta-analyses [@devries2024].
+Personalized luminous exposure data are progressively gaining importance across various domains, including research, occupational affairs, and lifestyle tracking. Data are collected through a increasing selection of wearable light loggers and dosimeters, varying in size, shape, functionality, and output format [@hartmeyer2023]. Despite or potentially because of numerous use cases, the field still lacks a unified framework for collecting, validating, and analyzing the accumulated data [@hartmeyer2023; @spitschan2022].  This issue increases the time and expertise necessary to handle such data and also compromises the FAIRness (findability, accessibility, interoperability, reusability) [@wilkinson2016] of the results, especially for meta-analyses [@devries2024].
 
-`LightLogR` (\autoref{fig:one}) was designed to be used by researchers who deal with personal light exposure data collected from wearable devices. These data are of interest for various disciplines, including chronobiology, sleep research, vision science and epidemiology, as well as for post-occupancy evaluations in  architecture and lighting design. The package is intended to streamline the process of importing,  processing, and analysing these data in a reproducible and transparent manner. The package is available on GitHub (@repo) and CRAN (@cran), has a dedicated website for documentation and tutorials (@documentation), and releases are archived on Zenodo (@zenodo).
+`LightLogR` was designed to be used by researchers who deal with personal light exposure data collected from wearable devices (\autoref{fig:one}). These data are of interest for various disciplines, including chronobiology, sleep research, vision science and epidemiology, as well as for post-occupancy evaluations in  architecture and lighting design. The package is intended to streamline the process of importing,  processing, and analysing these data in a reproducible and transparent manner. The package is available on GitHub (@repo) and CRAN (@cran), has a dedicated website for documentation and tutorials (@documentation), and releases are archived on Zenodo (@zenodo).
 
 `LightLogR`'s key features include:
 
@@ -68,9 +68,9 @@ Personalized luminous exposure data is progressively gaining importance across v
 
 -   preprocessing functions to combine different time series, aggregate and filter data, and find and deal with implicitly missing data,
 
--   visualization functions to quickly explore the data. These function are based on the popular `ggplot2` [@wickham2016] plotting package and are designed to be easily customizable to construct publication-ready figures (see, \autoref{fig:two}),
[…]
