### 2025-02-11 7957543
created the number_states function
