+++ NAMESPACE
@@ -35,6 +35,7 @@ export(gg_day)
 export(gg_days)
 export(gg_doubleplot)
 export(gg_overview)
+export(gg_photoperiod)
 export(import)
 export(import_Dataset)
 export(import_Statechanges)
@@ -45,6 +46,7 @@ export(intradaily_variability)
 export(join_datasets)
 export(ll_import_expr)
 export(midpointCE)
+export(number_states)
 export(nvRC)
 export(nvRC_circadianBias)
 export(nvRC_circadianDisturbance)
