### 2025-02-11 0d71304
created the gg_photoperiod function
