### 2025-02-11 7586a5d
Increment version number to 0.5.0
