### 2025-02-12 b02a110
Update to photoperiod functions
