### 2025-02-12 1c37394
corrected a bug that left POSIXct columns out of aggregations
