+++ man/aggregate_Datetime.Rd
@@ -13,6 +13,7 @@ aggregate_Datetime(
   character.handler = function(x) names(which.max(table(x, useNA = "ifany"))),
   logical.handler = function(x) mean(x) >= 0.5,
   factor.handler = function(x) factor(names(which.max(table(x, useNA = "ifany")))),
+  datetime.handler = mean,
   ...
 )
 }
@@ -35,8 +36,8 @@ data at all.}
 \item{type}{One of \code{"round"}(the default), \code{"ceiling"} or \code{"floor"}. Setting
 chooses the relevant function from \pkg{lubridate}.}
 
-\item{numeric.handler, character.handler, logical.handler, factor.handler}{functions that handle the respective data types. The default handlers
-calculate the \code{mean} for \code{numeric} and the \code{mode} for \code{character}, \code{factor}
+\item{numeric.handler, character.handler, logical.handler, factor.handler, datetime.handler}{functions that handle the respective data types. The default handlers
+calculate the \code{mean} for \code{numeric} and \code{POSIXct} and the \code{mode} for \code{character}, \code{factor}
 and \code{logical} types.}
 
 \item{...}{arguments given over to \code{\link[dplyr:summarise]{dplyr::summarize()}} to handle columns
