### 2025-02-12 a1c817a
adding rmarkdown to requirements
