+++ DESCRIPTION
@@ -63,3 +63,4 @@ Suggests:
     testthat (>= 3.0.0),
     tidyverse
 Config/testthat/edition: 3
+Config/Needs/website: rmarkdown
