+++ pkgdown.yml
@@ -5,8 +5,9 @@ articles:
   articles/Day: Day.html
   articles/Import: Import.html
   articles/Metrics: Metrics.html
+  articles/photoperiod: photoperiod.html
   articles/Visualizations: Visualizations.html
-last_built: 2025-01-23T11:21Z
+last_built: 2025-02-12T17:18Z
 urls:
   reference: https://tscnlab.github.io/LightLogR/reference
   article: https://tscnlab.github.io/LightLogR/articles
