+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   articles/Metrics: Metrics.html
   articles/photoperiod: photoperiod.html
   articles/Visualizations: Visualizations.html
-last_built: 2025-02-12T17:18Z
+last_built: 2025-02-12T19:37Z
 urls:
   reference: https://tscnlab.github.io/LightLogR/reference
   article: https://tscnlab.github.io/LightLogR/articles
