### 2025-02-12 #7601 @JZauner
Hi @rwegener2! We have updated the [main branch](https://github.com/tscnlab/LightLogR) to include the changes in the paper. Your suggestion on the reference in the table solved the issue! Because some preprints we mentioned in the paper are now published, we changed their references to point to the final publications. Sorry if that requires a new check for the doi's.
