### 2025-02-12 #7601 @JZauner
Also, we reference Table 1 in the text - it just did not show up because of the mislabeling.
