### 2025-02-13 #7601 @rwegener2
Thanks for the updates @JZauner! Those are all great additions.

It looks like this JOSS issue is pointing to a branch called `paper` to render the pdf. Would you mind adding your latest updates to the `paper` branch as well? If you want to switch the associated branch to `main` it's probably possible, but I will have to ask the JOSS dev team how to do that mid-review.
