### 2025-02-17 1d4de92
referencing photoperiod capabilities
