### 2025-02-17 8136249
news update for 0.5.1
