### 2025-02-18 779b0c2
Increment version number to 0.5.2
