### 2025-02-18 46b5174
included new function normalize_counts
