+++ NAMESPACE
@@ -46,6 +46,7 @@ export(intradaily_variability)
 export(join_datasets)
 export(ll_import_expr)
 export(midpointCE)
+export(normalize_counts)
 export(number_states)
 export(nvRC)
 export(nvRC_circadianBias)
