### 2025-02-18 f1b7785
inclusion of new functions and bindings
