+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   articles/Metrics: Metrics.html
   articles/photoperiod: photoperiod.html
   articles/Visualizations: Visualizations.html
-last_built: 2025-02-18T07:25Z
+last_built: 2025-02-18T13:16Z
 urls:
   reference: https://tscnlab.github.io/LightLogR/reference
   article: https://tscnlab.github.io/LightLogR/articles
