+++ NEWS.md
@@ -1,8 +1,10 @@
+# LightLogR 0.5.3
+
 # LightLogR 0.5.2
 
 * passed 300 unit tests for `LightLogR` 🎉
 
-* added the low-level helper function `normalize_counts()` and the accompanying dataset `gain.ratio.tables` to facilitate calculating normalized sensor values when comparing across different sensors, e.g. to assess daylighting conditions based on UV, IR, and photopic sensing ranges. See documentation for more infos.
+* `normalize_counts()` was added as a low-level helper function  and the accompanying dataset `gain.ratio.tables` to facilitate calculating normalized sensor values when comparing across different sensors, e.g. to assess daylighting conditions based on UV, IR, and photopic sensing ranges. See documentation for more infos.
 
 * Update to the import function of `GENEActiv` devices, based on input from the author of the GGIR package. The timezone `tz` argument in LightLogR now is just set on the timestamp provided by the GGIR export, instead of shifting the datetime. This requires the correct setting of `desiredtz`/`configtz` arguments in GGIR during preprocessing.
 
@@ -42,7 +44,7 @@
 
 * added support for LIMO devices
 
-* added support for GENEActiv devices, when data was preprocessed with the [`GGIR`](https://cran.r-project.org/web/packages/GGIR/vignettes/GGIR.html) package. The function `import$GENEActiv_GGIR()` takes the `GGIR` output and imports it with LightLogR naming schemes. #27
+* added support for GENEActiv devices, when data was preprocessed with the [`GGIR`]( https://cran.r-project.org/package=GGIR) package. The function `import$GENEActiv_GGIR()` takes the `GGIR` output and imports it with LightLogR naming schemes. #27
 
 # LightLogR 0.4.0 "Nautical dawn"
 
