### 2025-02-20 2e7888c
small cleanup changes for CRAN submission
