### 2025-02-20 #7601 @JZauner
Great! Thanks @rwegener2 !

The final version number is v0.5.3 (Civil Dawn) and the DOI to the Zenodo Archive repo is 10.5281/zenodo.14899662 (https://doi.org/10.5281/zenodo.14899662)
