### 2025-02-20 #7601 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
