### 2025-02-21 #7601 @Kevin-Mattheus-Moerman
- [ ] Please ensure the archive title matches the paper title: `LightLogR: Reproducible analysis of personal light
exposure data`
- [ ] Words like `analysing` and `programme` are in European English, while words like `standardizes`, `visualization`, `Personalized`, and `customizable` are in American English. We recommend you use one style consistently.
