+++ paper/paper.md
@@ -52,13 +52,13 @@ affiliations:
 
 # Summary
 
-Light plays an important role in human health and well-being, which necessitates the study of the effects of personal light exposure in real-world settings, measured by means of wearable devices. A growing number of studies incorporate these kinds of data to assess associations between light and health outcomes. Yet with few or missing standards, guidelines, and frameworks, it is challenging setting up measurements, analysing the data, and comparing outcomes between studies. Overall, time series data from wearable light loggers are significantly more complex compared to controlled stimuli used in laboratory studies. In this paper, we introduce `LightLogR`, a novel resource to facilitate these research efforts. The package for R statistical software is open-source and permissively MIT-licenced. As part of a developing software ecosystem, `LightLogR` is built with common challenges of current and future datasets in mind. The package standardizes many tasks for importing and processing personal light exposure data. It allows for quick as well as detailed insights into the datasets through summary and visualization tools. Furthermore, LightLogR incorporates major metrics commonly used in the field (61 metrics across 17 metric families), all while embracing an inherently hierarchical, participant-based data structure.
+Light plays an important role in human health and well-being, which necessitates the study of the effects of personal light exposure in real-world settings, measured by means of wearable devices. A growing number of studies incorporate these kinds of data to assess associations between light and health outcomes. Yet with few or missing standards, guidelines, and frameworks, it is challenging setting up measurements, analysing the data, and comparing outcomes between studies. Overall, time series data from wearable light loggers are significantly more complex compared to controlled stimuli used in laboratory studies. In this paper, we introduce `LightLogR`, a novel resource to facilitate these research efforts. The package for R statistical software is open-source and permissively MIT-licenced. As part of a developing software ecosystem, `LightLogR` is built with common challenges of current and future datasets in mind. The package standardises many tasks for importing and processing personal light exposure data. It allows for quick as well as detailed insights into the datasets through summary and visualisation tools. Furthermore, LightLogR incorporates major metrics commonly used in the field (61 metrics across 17 metric families), all while embracing an inherently hierarchical, participant-based data structure.
 
 ![LightLogR logo \label{fig:one}](logo.png){width="25%"}
 
 # Statement of need
 
-Personalized luminous exposure data are progressively gaining importance across various domains, including research, occupational affairs, and lifestyle tracking. Data are collected through a increasing selection of wearable light loggers and dosimeters, varying in size, shape, functionality, and output format [@hartmeyer2023]. Despite or potentially because of numerous use cases, the field still lacks a unified framework for collecting, validating, and analyzing the accumulated data [@hartmeyer2023; @spitschan2022].  This issue increases the time and expertise necessary to handle such data and also compromises the FAIRness (findability, accessibility, interoperability, reusability) [@wilkinson2016] of the results, especially for meta-analyses [@devries2024].
+Personalised luminous exposure data are progressively gaining importance across various domains, including research, occupational affairs, and lifestyle tracking. Data are collected through a increasing selection of wearable light loggers and dosimeters, varying in size, shape, functionality, and output format [@hartmeyer2023]. Despite or potentially because of numerous use cases, the field still lacks a unified framework for collecting, validating, and analyzing the accumulated data [@hartmeyer2023; @spitschan2022].  This issue increases the time and expertise necessary to handle such data and also compromises the FAIRness (findability, accessibility, interoperability, reusability) [@wilkinson2016] of the results, especially for meta-analyses [@devries2024].
 
[…]
