### 2025-02-27 cdaa9a1
fixed error with non MEDI y-columns
