### 2025-02-27 4415ad8
better bug fix for inherited aesthetics
